(PC-24831)[PRO] refactor: fix typo in component name
