+++ pro/src/screens/IndividualOffer/IndividualOfferLayout/IndividualOfferLayout.module.scss
