+++ pro/src/screens/IndividualOffer/IndividualOfferLayout/OfferStatusBanner/OfferStatusBanner.tsx
