+++ pro/src/screens/IndividualOffer/IndividualOfferLayout/utils/__specs__/getTitle.spec.ts
