+++ pro/src/screens/IndividualOffer/IndividualOfferLayout/utils/getTitle.ts
